// swagger/generated-code/apis/GeographicalObjectApi.ts
export interface ApiGeographicalObjectCreatePostRequest {
    authorization: string;
    geographicalObjectSerializer: GeographicalObjectSerializer;
}
// ...
export interface ApiGeographicalObjectIdUpdateImagePutRequest {
    id: number;
    authorization: string;
    photo: Blob;
}